reddit-1/requirements.txt
# Function dependencies, for example:
# package>=version

requests==2.25.1
google-cloud-logging
google-cloud-storage
openai